backend/requirements.txt
Django==4.2.9
djangorestframework>=3.15.1,<3.16.0
django-cors-headers>=4.3.1,<5.0.0
psycopg2-binary>=2.9.9,<2.10.0
django-environ>=0.11.2,<0.12.0
gunicorn>=22.0.0,<23.0.0
whitenoise==6.6.0
python-dotenv==1.0.1
requests>=2.31.0,<3.0.0

# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-mock==3.12.0
pytest-cov==4.1.0

# Authentication
django-allauth>=0.56.0,<0.58.0
dj-rest-auth[with_social]==5.0.2
